feat: Navigation bar and router set up
