+++ package.json
@@ -46,6 +46,7 @@
     "@mui/lab": "^5.0.0-alpha.64",
     "@mui/material": "^5.2.8",
     "@mui/styled-engine-sc": "^5.1.0",
+    "antd": "^4.18.4",
     "d3": "^7.1.1",
     "d3-delaunay": "^6.0.2",
     "loadash": "^1.0.0",
@@ -55,7 +56,8 @@
     "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^6.0.1",
+    "react-router-dom": "^6.2.1",
+    "recoil": "^0.5.2",
     "save-svg-as-png": "^1.4.17",
     "semantic-ui-react": "^2.0.4",
     "split-polygon": "^1.0.0",
