+++ src/components/app/App.css
@@ -1,0 +1,6 @@
+.title {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding-top: 15px;
+}
