+++ src/features/About/About.css
@@ -1,0 +1,3 @@
+.about-page {
+  padding: 15px;
+}
