+++ src/features/About/About.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import './About.css';
+
+export default function About() {
+  return (
+    <div className="about-page">
+      <h2>About Waggle</h2>
+      <span>Waggle is a web application allowing you to visually program a simulated robot swarm.
+      Its primary purpose is to provide a gentle introduction to the field of Swarm Robotics.</span>
+    </div>
+  );
+}
