+++ src/features/SimulationDescription/SimulationDescription.css
@@ -1,0 +1,5 @@
+.simulations {
+  padding: 15px;
+  display: flex;
+  flex-direction: column;
+}
