+++ src/features/Tutorial.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function Tutorial() {
+  return (
+    <h2>Tutorial</h2>
+  );
+}
